+++ lib/common/widgets/appbar/appbar.dart
@@ -1,8 +1,11 @@
 import 'package:e_store/utils/constants/sizes.dart';
 import 'package:e_store/utils/device/device_utility.dart';
+import 'package:e_store/utils/helpers/helper_functions.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:iconsax/iconsax.dart';
+
+import '../../../utils/constants/colors.dart';
 
 
 class TAppBar extends StatelessWidget implements PreferredSizeWidget {
@@ -23,15 +26,15 @@ class TAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   @override
   Widget build(BuildContext context) {
+    final dark = THelperFunctions.isDarkMode(context);
     return Padding(
         padding: const EdgeInsets.symmetric(horizontal: TSizes.md),
         child: AppBar(
           automaticallyImplyLeading: false,
-          backgroundColor: Colors.transparent,
           leading: showBackArrow
               ? IconButton(
                   onPressed: () => Get.back(),
-                  icon: const Icon(Iconsax.arrow_left))
+                  icon: Icon(Iconsax.arrow_left, color: dark ? TColors.white : TColors.black,))
               : leadingIcon != null
                   ? IconButton(
                       onPressed: leadingOnPressed, icon: Icon(leadingIcon))
