+++ lib/common/widgets/products.cart/cart_menu_icon.dart
@@ -1,43 +1,1 @@
-import 'package:flutter/material.dart';
-import 'package:iconsax/iconsax.dart';
-
-import '../../../../utils/constants/colors.dart';
-
-
-class TCartCounterIcon extends StatelessWidget {
-  const TCartCounterIcon({
-    super.key,
-    this.iconColor,
-    required this.onPressed
-  });
-
-  final VoidCallback onPressed;
-  final Color? iconColor;
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        IconButton(
-            onPressed: onPressed,
-            icon: Icon(Iconsax.shopping_bag, color: iconColor)),
-        Positioned(
-            right: 0,
-            child: Container(
-              width: 18,
-              height: 18,
-              decoration: BoxDecoration(
-                color: TColors.black,
-                borderRadius: BorderRadius.circular(100),
-              ),
-              child: Center(
-                  child: Text("2",
-                      style: Theme.of(context)
-                          .textTheme
-                          .labelLarge!
-                          .apply(color: TColors.white, fontSizeFactor: 0.8))),
-            ))
-      ],
-    );
-  }
-}
+// TODO Implement this library.
