+++ lib/common/widgets/products/product_cards/product_card_vertical.dart
@@ -1,0 +1,139 @@
+import 'package:e_store/utils/constants/colors.dart';
+import 'package:e_store/utils/helpers/helper_functions.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/painting.dart';
+import 'package:iconsax/iconsax.dart';
+
+import '../../../../rounded_container.dart';
+import '../../../../utils/constants/image_strings.dart';
+import '../../../../utils/constants/sizes.dart';
+import '../../../styles/shadows.dart';
+import '../../icons/t_circular_icon.dart';
+import '../../images/t_rounded_image.dart';
+import '../../texts/product_price_text.dart';
+import '../../texts/product_title_text.dart';
+
+class TProductCardVertical extends StatelessWidget {
+  const TProductCardVertical({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final dark = THelperFunctions.isDarkMode(context);
+    return GestureDetector(
+      onTap: () {},
+      child: Container(
+        width: 180,
+        padding: const EdgeInsets.all(1),
+        decoration: BoxDecoration(
+          boxShadow: [TShadowStyle.verticalProductShadow],
+          borderRadius: BorderRadius.circular(TSizes.productImageRadius),
+          color: dark ? TColors.darkerGrey : TColors.white,
+        ),
+        child: Column(
+          children: [
+            TRoundedContainer(
+              height: 180,
+              padding: const EdgeInsets.all(TSizes.sm),
+              backgroundColor: dark ? TColors.dark : TColors.light,
+              child: Stack(
+                children: [
+                  const TRoundedImage(
+                    imageUrl: TImages.productImage1,
+                    applyImageRadius: true,
+                  ),
+                  Positioned(
+                    top: 12,
+                    child: TRoundedContainer(
+                      radius: TSizes.sm,
+                      backgroundColor: TColors.secondary.withOpacity(0.8),
+                      padding: const EdgeInsets.symmetric(
+                          horizontal: TSizes.sm, vertical: TSizes.xs),
+                      child: Text(
+                        '25%',
+                        style: Theme.of(context)
+                            .textTheme
+                            .labelLarge
+                            ?.apply(color: TColors.black),
+                      ),
+                    ),
+                  ),
+                  const Positioned(
+                      top: 0,
+                      right: 0,
+                      child: TCircularIcon(
+                        icon: Iconsax.heart,
+                        color: Colors.red,
+                      ))
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(left: TSizes.sm),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const TProductTitleText(
+                    title: 'Green Nike Air Shoes',
+                    smallSize: true,
+                  ),
+                  const SizedBox(
+                    height: TSizes.spaceBtwItems / 2,
+                  ),
+                  Row(
+                    children: [
+                      Text(
+                        'Nike',
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 1,
+                        style: Theme.of(context).textTheme.labelMedium,
+                      ),
+                      const SizedBox(
+                        width: TSizes.xs,
+                      ),
+                      const Icon(
+                        Iconsax.verify5,
+                        color: TColors.primary,
+                        size: TSizes.iconXs,
+                      )
+                    ],
+                  ),
+                ],
+              ),
+            ),
+
+            // Todo: Add Spacer() here to keep height of each Box same in case 1 or case 2 lines of Headings
+            const Spacer(),
+
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                ///Price
+                const Padding(
+                    padding: EdgeInsets.only(left: TSizes.sm),
+                    child: TProductPriceText(
+                      price: '35.5',
+                    )),
+
+                ///Add to cart
+                Container(
+                  decoration: const BoxDecoration(
+                      color: TColors.dark,
+                      borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(TSizes.cardRadiusMd),
+                          bottomRight:
+                              Radius.circular(TSizes.productImageRadius))),
+                  child: const SizedBox(
+                      width: TSizes.iconLg * 1.2,
+                      height: TSizes.iconLg * 1.2,
+                      child: Center(
+                          child: Icon(Iconsax.add, color: TColors.white))),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
