+++ lib/common/widgets/products/ratings/rating_indicator.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_rating_bar/flutter_rating_bar.dart';
+import 'package:iconsax/iconsax.dart';
+
+import '../../../../utils/constants/colors.dart';
+
+class TRatingBarIndicator extends StatelessWidget {
+  const TRatingBarIndicator({
+    super.key,
+    required this.rating,
+  });
+
+  final double rating;
+  @override
+  Widget build(BuildContext context) {
+    return RatingBarIndicator(
+        rating: rating,
+        itemSize: 20,
+        unratedColor: TColors.grey,
+        itemBuilder: (_, __) => const Icon(
+              Iconsax.star,
+              color: TColors.primary,
+            ));
+  }
+}
