+++ lib/features/shop/screens/home/home.dart
@@ -1,4 +1,3 @@
-import 'package:e_store/common/widgets/products.cart/product_cards/product_card_vertical.dart';
 import 'package:e_store/features/shop/screens/home/widgets/home_appbar.dart';
 import 'package:e_store/features/shop/screens/home/widgets/promo_slider.dart';
 import 'package:e_store/section_heading.dart';
@@ -9,6 +8,7 @@ import 'package:flutter/material.dart';
 import '../../../../common/widgets/custom_shapes/containers/primary_header_container.dart';
 import '../../../../common/widgets/custom_shapes/containers/search_container.dart';
 import '../../../../common/widgets/layouts/grid_layout.dart';
+import '../../../../common/widgets/products/product_cards/product_card_vertical.dart';
 import '../../../../home_categories.dart';
 import '../../../../utils/constants/colors.dart';
