+++ lib/features/shop/screens/home/widgets/home_appbar.dart
@@ -1,7 +1,7 @@
 import 'package:e_store/common/widgets/appbar/appbar.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../../common/widgets/products.cart/cart_menu_icon.dart';
+import '../../../../../common/widgets/products/cart_menu_icon.dart';
 import '../../../../../utils/constants/colors.dart';
 import '../../../../../utils/constants/text_strings.dart';
