+++ lib/features/shop/screens/porduct_reviews/widgets/progess_indicator_and_rating.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+import '../../../../../utils/constants/colors.dart';
+import '../../../../../utils/device/device_utility.dart';
+
+class TRatingProgressIndicator extends StatelessWidget {
+  const TRatingProgressIndicator({
+    super.key,
+    required this.text,
+    required this.value,
+  });
+
+  final String text;
+  final double value;
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Expanded(flex: 1,child: Text('5', style: Theme.of(context).textTheme.bodyMedium)),
+        Expanded(
+          flex: 11,
+          child: SizedBox(
+            width: TDeviceUtils.getScreenWidth(context) * 0.8,
+            child: LinearProgressIndicator(
+              value: value,
+              minHeight: 25,
+              backgroundColor: TColors.grey,
+              borderRadius: BorderRadius.circular(7),
+              valueColor: const AlwaysStoppedAnimation(TColors.primary),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
