+++ lib/features/shop/screens/porduct_reviews/widgets/user_review_cart.dart
@@ -1,0 +1,84 @@
+import 'package:e_store/common/widgets/products/ratings/rating_indicator.dart';
+import 'package:e_store/rounded_container.dart';
+import 'package:e_store/utils/constants/colors.dart';
+import 'package:e_store/utils/constants/image_strings.dart';
+import 'package:e_store/utils/constants/sizes.dart';
+import 'package:e_store/utils/helpers/helper_functions.dart';
+import 'package:flutter/material.dart';
+import 'package:readmore/readmore.dart';
+class UserReviewCard extends StatelessWidget {
+  const UserReviewCard({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final dark = THelperFunctions.isDarkMode(context);
+    return Column(
+      children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Row(
+              children: [
+                const CircleAvatar(backgroundImage: AssetImage(TImages.userProfileImage1)),
+                const SizedBox(width: TSizes.spaceBtwItems,),
+                Text('Jonny Doe', style: Theme.of(context).textTheme.titleLarge),
+              ],
+            ),
+            IconButton(onPressed: () {}, icon: const Icon(Icons.more_vert)),
+          ],
+        ),
+        const SizedBox(height: TSizes.spaceBtwItems),
+
+        /// Review
+        Row(
+          children: [
+            const TRatingBarIndicator(rating: 4),
+            const SizedBox(width: TSizes.spaceBtwItems),
+            Text('01 Nov, 2023', style: Theme.of(context).textTheme.bodyMedium),
+          ],
+        ),
+        const SizedBox(width: TSizes.spaceBtwItems),
+        const ReadMoreText(
+          'The user interface of the app is quite intuitive. I was able to navigate and make purchase seamlessly. Great job!',
+          trimLines: 2,
+          trimMode: TrimMode.Line,
+          trimExpandedText: 'show less',
+          trimCollapsedText: 'show more',
+          moreStyle: TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: TColors.primary),
+          lessStyle: TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: TColors.primary),
+        ),
+        const SizedBox(width: TSizes.spaceBtwItems),
+
+        /// Company Review
+        TRoundedContainer(
+          backgroundColor: dark ? TColors.darkerGrey : TColors.grey,
+          child: Padding(
+            padding: EdgeInsets.all(TSizes.md),
+            child: Column(
+              children: [
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text('eShop', style: Theme.of(context).textTheme.titleMedium),
+                    Text('01 Nov, 2023', style: Theme.of(context).textTheme.bodyMedium),
+                  ],
+                ),
+                const SizedBox(width: TSizes.spaceBtwItems),
+                const ReadMoreText(
+                  'The user interface of the app is quite intuitive. I was able to navigate and make purchase seamlessly. Great job!',
+                  trimLines: 2,
+                  trimMode: TrimMode.Line,
+                  trimExpandedText: 'show less',
+                  trimCollapsedText: 'show more',
+                  moreStyle: TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: TColors.primary),
+                  lessStyle: TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: TColors.primary),
+                ),
+              ],
+            ),
+          ),
+        ),
+        const SizedBox(height: TSizes.spaceBtwSections,)
+      ],
+    );
+  }
+}
