+++ lib/features/shop/screens/product_details/product_details.dart
@@ -1,0 +1,7 @@
+import 'package:e_store/features/shop/screens/porduct_reviews/product_reviews.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get_core/src/get_main.dart';
+import 'package:iconsax/iconsax.dart';
+
+/// - Reviews
+// IconButton(icon: const Icon(Iconsax.arrow_right_3, size: 18), onPressed: () => Get.to(() => const ProductReviewsScreen()));
