+++ pubspec.yaml
@@ -24,8 +24,15 @@ dependencies:
   get_storage: ^2.1.1
   flutter_native_splash: ^2.4.0
   smooth_page_indicator: ^1.1.0
-  carousel_slider: ^4.2.1
 
+  # Icons
+
+  # Product Specific
+  readmore: ^2.2.0
+  carousel_slider: ^4.2.1
+  flutter_rating_bar: ^4.0.1
+
+#---------------- PACKAGE END----------------#
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -39,11 +46,12 @@ flutter:
   assets:
     - assets/logos/
     - assets/icons/
+    - assets/icons/categories/
     - assets/images/animations/
     - assets/images/on_boarding_images/
     - assets/images/banners/
     - assets/images/products/
-    - assets/icons/categories/
+    - assets/images/reviews/
 
   fonts:
     - family: Poppins
